components/Docs.tsx を新規作成
docs/page.tsx にコンポネントのみ設置する
